Add files via upload
The exe and the python file will run without error only if the img and sounds folders are in the same directory with all the files as in "Space Slayer Python Project.rar". 
Extract the rar file in any location and run the exe inside the folder in order to run without error.